backend/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
scipy
python-multipart
pydantic
python-dotenv
httpx
huggingface-hub
requests
google-generativeai
supabase
slack-sdk
discord.py
